pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "OpticalPooledScreens_DL"
description = "General DL repository for OPS work"
readme = "README.md"
requires-python = ">=3.7"
license = { text = "BSD 3-Clause License" }
authors = [
    { name = "DLMBL", email = "ryan.savill@web.de" }
]
classifiers = [
    "Programming Language :: Python :: 3",
]
dynamic = ["version"]
dependencies = [
    "pandas",
    "tifffile",
    "torch",
    "torchvision",
    "scikit-image",
    "matplotlib",
    "pathlib",
    "numpy",
    "zarr",
    "pyyaml",
    "scikit-learn",
    "seaborn",
    "piq",
    "tensorboard",
    "graphviz",
    "torchview",
]

[project.optional-dependencies]
dev = [
    "tqdm",
    "pytest",
]